refactor: split cli core e2e tests (#4939)
### Description

Splits cli core e2e tests to reduce test execution and reduces constants
duplicity across test files

### Drive-by changes


### Related issues


### Backward compatibility

- Yes

### Testing

- Manual/e2e
