+++ .github/workflows/test.yml
@@ -76,7 +76,7 @@ jobs:
             .yarn
           key: ${{ runner.os }}-yarn-4.5.1-cache-${{ hashFiles('./yarn.lock') }}
           fail-on-cache-miss: true
-      
+
       # Build required before linting or the intra-monorepo package cycle checking won't work
       - name: yarn-build
         uses: ./.github/actions/yarn-build-with-cache
@@ -126,7 +126,9 @@ jobs:
       fail-fast: false
       matrix:
         test:
-          - core
+          - core-apply
+          - core-deploy
+          - core-read
           - relay
           - warp-read
           - warp-apply
